requirement.txt
selenium==4.7.2
appium-python-client == 2.7.1
webdriver-manager==3.8.5

behave == 1.2.6
allure-behave==2.12.0

pdfplumber==0.7.6
openpyxl==3.0.10
dictdiffer==0.9.0
Pillow==9.3.0
